requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
Brotli==1.0.9
ccsm==0.9.14.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.3
colorlog==6.6.0
command-not-found==0.3
compizconfig-python==0.9.14.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flake8==6.0.0
frozenlist==1.3.0
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.7.0
monotonic==1.5
multidict==6.0.2
mutagen==1.45.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
py-cord==2.0.0
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.15.0
pycups==1.9.73
pyflakes==3.0.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
websockets==10.3
xkit==0.0.0
yarl==1.7.2
yt-dlp==2022.7.18
